requirements.txt
opencv-python
mediapipe
pandas
numpy
tqdm
scikit-image
matplotlib
pillow
